+++ src/main/java/com/chessgame/Piece.java
@@ -1,25 +1,25 @@
-package com.example;
+package com.chessgame;
 
-public abstract class Piece {
+public class Piece {
     protected Position position;
     protected PieceColor color;
-  
+
     public Piece(PieceColor color, Position position) {
-        this.color = color;
-        this.position = position;
+      this.color = color;
+      this.position = position;
     }
-  
+
     public PieceColor getColor() {
         return color;
     }
-  
+
     public Position getPosition() {
         return position;
     }
-  
+
     public void setPosition(Position position) {
         this.position = position;
     }
-  
+
     public abstract boolean isValidMove(Position newPosition, Piece[][] board);
 }
