+++ src/main/java/com/chessgame/PieceColor.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.chessgame;
 
 public enum PieceColor {
     BLACK, WHITE;
