+++ src/main/java/com/chessgame/Position.java
@@ -1,18 +1,18 @@
-package com.example;
+package com.chessgame;
 
 public class Position {
     private int row;
     private int column;
-  
-    public Position(int i, int j) {
-        this.row = i;
-        this.column = j;
+
+    public Position(int row, int column) {
+        this.row = row;
+        this.column = column;
     }
-  
+
     public int getRow() {
         return row;
     }
-  
+
     public int getColumn() {
         return column;
     }
