--- src/main/java/com/example/Test.java
@@ -1,7 +1,0 @@
-package com.example;
-
-public class Test {
-    public static void main (String[] args) {
-        System.out.println("Hello, World !");
-    }
-}
